Updated Repository to fit ASP.NET 2.0
With the release of ASP.NET 2.0 while the course only needed very minor
tweaks to address changes, this repo needed significant alterations.
Primarily where we had updated this project to use the `xproj` file
format that replaced `project.json` we've replaced the `xproj` with the
`csproj` file format. Changes also needed to be make to the `Program.cs`
and `Startup.cs` files.